fix: ErrorCode 컴포넌트에 쓰이는 스타일 파일 분리
