+++ src/views/ErrorPage/components/ErrorCode.tsx
@@ -1,6 +1,15 @@
 import { useIsMobile } from '@src/hooks/useDevice';
-import { Ic404Back, Ic404Front, Ic404Ghost, Ic404GhostDark, Ic500Back, Ic500Cone, Ic500ConeDark, Ic500Front } from '../assets';
-import * as S from '../styles';
+import {
+  Ic404Back,
+  Ic404Front,
+  Ic404Ghost,
+  Ic404GhostDark,
+  Ic500Back,
+  Ic500Cone,
+  Ic500ConeDark,
+  Ic500Front,
+} from '../assets';
+import * as S from './ErrorCode.style';
 
 interface ErrorCodeProps {
   code: 404 | 500;
@@ -23,41 +32,45 @@ export default function ErrorCode({ code }: ErrorCodeProps) {
   };
 
   const animationProps = {
-    initial : 'initial',
-    whileHover : 'whileHover',
-    transition : { duration: 0.3 },
+    initial: 'initial',
+    whileHover: 'whileHover',
+    transition: { duration: 0.3 },
   };
 
   return (
     <S.ErrorCode {...animationProps} variants={codeVariant}>
-    {code === 404 ? (
-      <>
-        {!isMobile &&
-          <>
-            <Ic404Front height={SIZE.height} />
-            <Ic404Back height={SIZE.height} />
-          </>
-        }
-        {isMobile ? <Ic404GhostDark /> :
-          <S.ErrorIcon {...animationProps} variants={iconVariant}>
-            <Ic404Ghost height={SIZE.icon} />
-          </S.ErrorIcon>
-        }
+      {code === 404 ? (
+        <>
+          {!isMobile && (
+            <>
+              <Ic404Front height={SIZE.height} />
+              <Ic404Back height={SIZE.height} />
+            </>
+          )}
+          {isMobile ? (
+            <Ic404GhostDark />
+          ) : (
+            <S.ErrorIcon {...animationProps} variants={iconVariant}>
+              <Ic404Ghost height={SIZE.icon} />
+            </S.ErrorIcon>
+          )}
         </>
       ) : (
-      <>
-        {!isMobile &&
-          <>
-            <Ic500Front height={SIZE.height} />
-            <Ic500Back height={SIZE.height} />
-          </>
-        }
-        {isMobile ? <Ic500ConeDark /> :
-          <S.ErrorIcon {...animationProps} variants={iconVariant}>
-            <Ic500Cone height={SIZE.icon} />
-          </S.ErrorIcon>
-        }
-      </>
+        <>
+          {!isMobile && (
+            <>
+              <Ic500Front height={SIZE.height} />
+              <Ic500Back height={SIZE.height} />
+            </>
+          )}
+          {isMobile ? (
+            <Ic500ConeDark />
+          ) : (
+            <S.ErrorIcon {...animationProps} variants={iconVariant}>
+              <Ic500Cone height={SIZE.icon} />
+            </S.ErrorIcon>
+          )}
+        </>
       )}
     </S.ErrorCode>
   );
